Notifications and first set of tests

+++ SwarmOS/src/SwarmManagers/TestInterruptManager.ts
@@ -1,38 +1,42 @@
 declare interface ITestData_Memory {
-    testComplete: boolean;
+    NotificationSent: number;
     SleeperReady: boolean;
     expectedSleep: number;
     begun: number;
+    testComplete: boolean
+    CurStep: number
 }
 
 declare var Memory: {
     TestData: ITestData_Memory
 }
+
 if (!Memory.TestData) {
     Memory.TestData = {
-        testComplete: false,
+        NotificationSent: 0,
         SleeperReady: false,
         expectedSleep: 0,
-        begun: 0
+        begun: 0,
+        CurStep: 0,
+        testComplete: false,
     }
 }
 
 import { ProcessBase, ExtensionBase } from "Core/BasicTypes";
 export const bundle: IPackage<ITestData_Memory> = {
     install(processRegistry: IProcessRegistry, extensionRegistry: IExtensionRegistry) {
-        processRegistry.register(PKG_TestInterrupt, TestSleeper);
-        processRegistry.register('PKG_Test2', TestInterrupter);
+        processRegistry.register(FT_InterruptListener, TestSleeper);
+        processRegistry.register(FT_InterruptNotifier, TestInterrupter);
     },
-    rootImageName: PKG_TestInterrupt
+    rootImageName: FT_InterruptListener
 }
 const PKG_TestInterrupt_LogContext: LogContext = {
-    logID: PKG_TestInterrupt,
+    logID: FT_InterruptListener,
     logLevel: LOG_DEBUG
 }
 
 class TestSleeper extends ProcessBase {
-    protected OnLoad(): void {
-    }
+    protected OnLoad(): void { }
     constructor(protected context: IProcessContext) {
         super(context);
         Logger.CreateLogContext(PKG_TestInterrupt_LogContext);
@@ -44,13 +48,13 @@ class TestSleeper extends ProcessBase {
         return this._logger;
     }
     private _logger: ILogger = {
-        alert: (message: (string | (() => string))) => { Logger.alert(message, PKG_TestInterrupt); },
-        debug: (message: (string | (() => string))) => { Logger.debug(message, PKG_TestInterrupt); },
-        error: (message: (string | (() => string))) => { Logger.error(message, PKG_TestInterrupt); },
-        fatal: (message: (string | (() => string))) => { Logger.fatal(message, PKG_TestInterrupt); },
-        info: (message: (string | (() => string))) => { Logger.info(message, PKG_TestInterrupt); },
-        trace: (message: (string | (() => string))) => { Logger.trace(message, PKG_TestInterrupt); },
-        warn: (message: (string | (() => string))) => { Logger.warn(message, PKG_TestInterrupt); },
+        alert: (message: (string | (() => string))) => { Logger.alert(message, FT_InterruptListener); },
+        debug: (message: (string | (() => string))) => { Logger.debug(message, FT_InterruptListener); },
+        error: (message: (string | (() => string))) => { Logger.error(message, FT_InterruptListener); },
+        fatal: (message: (string | (() => string))) => { Logger.fatal(message, FT_InterruptListener); },
+        info: (message: (string | (() => string))) => { Logger.info(message, FT_InterruptListener); },
+        trace: (message: (string | (() => string))) => { Logger.trace(message, FT_InterruptListener); },
+        warn: (message: (string | (() => string))) => { Logger.warn(message, FT_InterruptListener); },
         CreateLogContext: Logger.CreateLogContext,
         DumpLogToConsole: Logger.DumpLogToConsole
     }
@@ -58,33 +62,71 @@ class TestSleeper extends ProcessBase {
     handleMissingMemory() {
         if (!Memory.TestData) {
             Memory.TestData = {
-                testComplete: false,
+                NotificationSent: 0,
                 SleeperReady: false,
                 expectedSleep: 0,
-                begun: 0
+                begun: 0,
+                CurStep: 0,
+                testComplete: false
             }
         }
         return Memory.TestData;
     }
+
     executeProcess(): void {
-        this.log.debug('Sleeper tick');
-        if (!this.memory.SleeperReady) {
-            this.log.debug('Set subscription')
-            this.interupter.Subscribe('test', this.pid);
-            this.memory.SleeperReady = true;
-            let randomSleep = GetRandomIndex(primes_300);
-            this.memory.expectedSleep = Game.time + randomSleep;
-            this.memory.begun = Game.time;
-            this.log.info(`Sleep for ${randomSleep} -- Expected wake up @ ${this.memory.expectedSleep}`)
-            this.SetProcessToSleep(randomSleep);
-        } else {
-            this.log.debug(`Retick occurred @ ${Game.time}`);
-            this.kernel.killProcess(this.pid);
-            this.memory.testComplete = false;
-            this.memory.SleeperReady = false;
-            this.memory.expectedSleep = 0;
-            this.memory.begun = 0;
+        if (this.memory.SleeperReady && this.memory.NotificationSent) {
+            if (this.memory.expectedSleep != Game.time || this.memory.testComplete) {
+                if (!this.memory.testComplete) {
+                    this.log.error(`Test has failed on step: ${this.memory.CurStep}`);
+                }
+                this.kernel.killProcess(this.pid);
+                this.memory.testComplete = false;
+                this.memory.NotificationSent = 0;
+                this.memory.SleeperReady = false;
+                this.memory.expectedSleep = 0;
+                this.memory.begun = 0;
+                this.memory.CurStep = 0;
+                return;
+            } else {
+                this.log.info(`Test has passed step: ${this.memory.CurStep}`)
+                this.memory.CurStep++;
+                this.memory.NotificationSent = 0;
+            }
         }
+
+        this.log.debug(`Sleeper tick: ${this.memory.CurStep}`);
+        switch (this.memory.CurStep) {
+            case (0):
+                this.interrupter.Subscribe('Test0', this.pid);
+                this.interrupter.Subscribe('Test1', this.pid);
+                this.interrupter.Subscribe('Test2', this.pid);
+                this.memory.expectedSleep = Game.time + 5;
+                break;
+            case (1):
+                this.interrupter.UnSubscribe(`Test0`, this.pid);
+                this.memory.expectedSleep = Game.time + 10;
+                break;
+            case (2):
+                this.interrupter.Subscribe('Test0', this.pid);
+                this.memory.expectedSleep = Game.time + 5;
+                break;
+            case (3):
+                this.interrupter.UnSubscribe('Test1', this.pid);
+                this.memory.expectedSleep = Game.time + 5;
+                break;
+            case (4):
+                this.interrupter.UnSubscribe('Test0', this.pid);
+                this.memory.expectedSleep = Game.time + 15;
+                break;
+            default:
+                this.memory.testComplete = true;
+                this.log.warn(`Test has completed`);
+                return;
+        }
+
+        this.memory.SleeperReady = true;
+        this.memory.begun = Game.time;
+        this.SetProcessToSleep(60);
     }
 }
 
@@ -101,33 +143,25 @@ class TestInterrupter extends ProcessBase {
         return this._logger;
     }
     private _logger: ILogger = {
-        alert: (message: (string | (() => string))) => { Logger.alert(message, PKG_TestInterrupt); },
-        debug: (message: (string | (() => string))) => { Logger.debug(message, PKG_TestInterrupt); },
-        error: (message: (string | (() => string))) => { Logger.error(message, PKG_TestInterrupt); },
-        fatal: (message: (string | (() => string))) => { Logger.fatal(message, PKG_TestInterrupt); },
-        info: (message: (string | (() => string))) => { Logger.info(message, PKG_TestInterrupt); },
-        trace: (message: (string | (() => string))) => { Logger.trace(message, PKG_TestInterrupt); },
-        warn: (message: (string | (() => string))) => { Logger.warn(message, PKG_TestInterrupt); },
+        alert: (message: (string | (() => string))) => { Logger.alert(message, FT_InterruptListener); },
+        debug: (message: (string | (() => string))) => { Logger.debug(message, FT_InterruptListener); },
+        error: (message: (string | (() => string))) => { Logger.error(message, FT_InterruptListener); },
+        fatal: (message: (string | (() => string))) => { Logger.fatal(message, FT_InterruptListener); },
+        info: (message: (string | (() => string))) => { Logger.info(message, FT_InterruptListener); },
+        trace: (message: (string | (() => string))) => { Logger.trace(message, FT_InterruptListener); },
+        warn: (message: (string | (() => string))) => { Logger.warn(message, FT_InterruptListener); },
         CreateLogContext: Logger.CreateLogContext,
         DumpLogToConsole: Logger.DumpLogToConsole
     }
 
-    handleMissingMemory() {
-        if (!Memory.TestData) {
-            Memory.TestData = {
-                testComplete: false,
-                SleeperReady: false,
-                expectedSleep: 0,
-                begun: 0
-            }
-        }
-        return Memory.TestData;
-    }
     executeProcess(): void {
         if (this.memory.SleeperReady) {
-            if (Game.time - 5 > this.memory.begun) {
-                this.interupter.Notify('test');
-                this.log.debug(`Notifying test`);
+            if (Game.time - 4 >= this.memory.begun) {
+                let id = `Test${this.memory.NotificationSent}`;
+                this.interrupter.Notify(id);
+                this.log.debug(`Notifying ${id}`);
+                this.memory.NotificationSent++;
+                this.memory.begun += 5;
             }
         }
     }
